Рефакторинг+Тесты+async\await
