+++ SescTool/DataParser.Tests/packages.config
@@ -1,0 +1,7 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="AngleSharp" version="0.9.9" targetFramework="net461" />
+  <package id="Extreme.Net" version="1.1.0.0" targetFramework="net461" />
+  <package id="NUnit" version="3.5.0" targetFramework="net461" />
+  <package id="NUnit3TestAdapter" version="3.6.0" targetFramework="net461" />
+</packages>
